Move list registration entries query test into a specfic file
